Add backup rotation toggle

+++ pg_backup_rotated.sh
@@ -163,38 +163,43 @@ function perform_backups()
 	echo -e "\nAll database backups complete!"
 }
 
-# MONTHLY BACKUPS
-
-DAY_OF_MONTH=`date +%d`
-
-if [ $DAY_OF_MONTH -eq 1 ];
+if [ $ROTATE_BACKUPS = "yes" ]
 then
-	# Delete all expired monthly directories
-	find $BACKUP_DIR -maxdepth 1 -name "*-monthly" -exec rm -rf '{}' ';'
+    # MONTHLY BACKUPS
 
-	perform_backups "-monthly"
+    DAY_OF_MONTH=`date +%d`
 
-	exit 0;
+    if [ $DAY_OF_MONTH -eq 1 ];
+    then
+        # Delete all expired monthly directories
+        find $BACKUP_DIR -maxdepth 1 -name "*-monthly" -exec rm -rf '{}' ';'
+
+        perform_backups "-monthly"
+
+        exit 0;
+    fi
+
+    # WEEKLY BACKUPS
+
+    DAY_OF_WEEK=`date +%u` #1-7 (Monday-Sunday)
+    EXPIRED_DAYS=`expr $((($WEEKS_TO_KEEP * 7) + 1))`
+
+    if [ $DAY_OF_WEEK = $DAY_OF_WEEK_TO_KEEP ];
+    then
+        # Delete all expired weekly directories
+        find $BACKUP_DIR -maxdepth 1 -mtime +$EXPIRED_DAYS -name "*-weekly" -exec rm -rf '{}' ';'
+
+        perform_backups "-weekly"
+
+        exit 0;
+    fi
+
+    # DAILY BACKUPS
+
+    # Delete daily backups 7 days old or more
+    find $BACKUP_DIR -maxdepth 1 -mtime +$DAYS_TO_KEEP -name "*-daily" -exec rm -rf '{}' ';'
+
+    perform_backups "-daily"
+else
+    perform_backups ""
 fi
-
-# WEEKLY BACKUPS
-
-DAY_OF_WEEK=`date +%u` #1-7 (Monday-Sunday)
-EXPIRED_DAYS=`expr $((($WEEKS_TO_KEEP * 7) + 1))`
-
-if [ $DAY_OF_WEEK = $DAY_OF_WEEK_TO_KEEP ];
-then
-	# Delete all expired weekly directories
-	find $BACKUP_DIR -maxdepth 1 -mtime +$EXPIRED_DAYS -name "*-weekly" -exec rm -rf '{}' ';'
-
-	perform_backups "-weekly"
-
-	exit 0;
-fi
-
-# DAILY BACKUPS
-
-# Delete daily backups 7 days old or more
-find $BACKUP_DIR -maxdepth 1 -mtime +$DAYS_TO_KEEP -name "*-daily" -exec rm -rf '{}' ';'
-
-perform_backups "-daily"
